requirements.txt
aiohttp==3.7.4.post0
aioredis==1.3.1
amqp==5.0.6
asgiref==3.3.4
asn1crypto==1.4.0
async-timeout==3.0.1
attrs==21.2.0
autobahn==21.3.1
Automat==20.2.0
backports.entry-points-selectable==1.1.0
billiard==3.6.4.0
bitarray==1.2.2
BitCash==0.6.3
bitcoinlib==0.5.3
Brlapi==0.6.7
cached-property==1.5.2
celery==5.1.2
certifi==2020.12.5
cffi==1.14.5
channels==3.0.4
channels-redis==3.3.0
chardet==4.0.0
chrome-gnome-shell==0.0.0
click==7.1.2
click-didyoumean==0.0.3
click-plugins==1.1.1
click-repl==0.2.0
coincurve==15.0.0
constantly==15.1.0
cryptography==3.4.7
cupshelpers==1.0
cytoolz==0.11.0
daphne==3.0.2
dateparser==1.0.0
distlib==0.3.2
distro==1.3.0
distro-info==0.21
dj-database-url==0.5.0
Django==3.2.3
django-celery-beat==2.2.1
django-celery-results==2.2.0
django-ckeditor==6.0.0
django-crispy==1.8.1
django-crispy-forms==1.11.2
django-js-asset==1.2.2
django-modeltranslation==0.17.1
django-timezone-field==4.2.1
django-user-agents==0.4.0
entrypoints==0.3
eth-abi==2.1.1
eth-account==0.5.4
eth-hash==0.3.1
eth-keyfile==0.5.1
eth-keys==0.3.3
eth-rlp==0.2.1
eth-typing==2.2.2
eth-utils==1.10.0
fastecdsa==2.1.5
filelock==3.0.12
gunicorn==20.1.0
hexbytes==0.2.1
hiredis==2.0.0
httplib2==0.11.3
hyperlink==21.0.0
idna==2.10
importlib-metadata==4.6.1
incremental==21.3.0
keyring==17.1.1
keyrings.alt==3.1.1
kombu==5.1.0
louis==3.8.0
Mako==1.0.7
MarkupSafe==1.1.0
msgpack==1.0.2
multidict==5.1.0
numpy==1.19.5
olefile==0.46
parsimonious==0.8.1
pexpect==4.6.0
Pillow==5.4.1
pipenv==2021.5.29
platformdirs==2.1.0
prompt-toolkit==3.0.19
psycopg2-binary==2.8.6
pyaes==1.6.1
pyasn1==0.4.8
pyasn1-modules==0.2.8
pycairo==1.16.2
pycparser==2.20
pycrypto==2.6.1
pycryptodome==3.10.1
pycups==1.9.73
pycurl==7.43.0.2
PyGObject==3.30.4
pyOpenSSL==20.0.1
PyQRCode==1.2.1
PySimpleSOAP==1.16.2
pysmbc==1.0.15.6
pyTelegramBotAPI==3.7.9
python-apt==1.8.4.3
python-binance==1.0.10
python-crontab==2.5.1
python-dateutil==2.8.1
python-debian==0.1.35
python-debianbts==2.8.2
python-memcached==1.59
pytz==2021.1
pyxdg==0.25
qrcode==6.1
regex==2021.4.4
reportbug===7.5.3-deb10u1
reportlab==3.5.13
requests==2.25.1
rlp==2.0.1
scrypt==0.8.17
SecretStorage==2.3.1
service-identity==21.1.0
six==1.16.0
SQLAlchemy==1.3.22
sqlparse==0.4.1
toolz==0.11.1
Twisted==21.7.0
txaio==21.2.1
typing-extensions==3.10.0.0
tzlocal==2.1
ua-parser==0.10.0
ujson==4.0.2
unattended-upgrades==0.1
urllib3==1.26.5
user-agents==2.2.0
vine==5.0.0
virtualenv==20.6.0
virtualenv-clone==0.5.5
wcwidth==0.2.5
websockets==9.1
whitenoise==5.2.0
yarl==1.6.3
zipp==3.5.0
zope.interface==5.4.0
